Move logic/proposals/hooks/genesis-epoch.ts -> logic/genesis-epoch/genesis-epoch.ts
