+++ src/logic/genesis-epoch/genesis-epoch.ts
@@ -1,8 +1,8 @@
 import { useCallback } from 'react';
-import { useChainData } from '../../../chain-data';
-import { useApi3Pool, usePossibleChainDataUpdate } from '../../../contracts';
-import { notifications } from '../../../components/notifications';
-import { messages } from '../../../utils';
+import { useChainData } from '../../chain-data';
+import { useApi3Pool, usePossibleChainDataUpdate } from '../../contracts';
+import { notifications } from '../../components/notifications';
+import { messages } from '../../utils';
 import { go } from '@api3/promise-utils';
 
 /**
